requirements.txt
alembic==1.13.1
annotated-types==0.7.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.4.1
click==8.1.7
Deprecated==1.2.15
dominate==2.9.1
dotenv-python==0.0.1
elastic-transport==8.13.1
elasticsearch==8.14.0
feedparser==6.0.11
Flask==3.0.3
Flask-Bootstrap4==4.0.2
Flask-Limiter==3.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
limits==4.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pymemcache==4.0.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
rsa==4.9
sgmllib3k==1.0.0
SQLAlchemy==2.0.30
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.3
wrapt==1.17.1
WTForms==3.1.2
